Conflicts:
	realm/realm-library/src/main/cpp/io_realm_RealmFileUserStore.cpp
	realm/realm-library/src/main/cpp/io_realm_SyncManager.cpp
	realm/realm-library/src/main/cpp/io_realm_internal_SharedRealm.cpp
	realm/realm-library/src/main/cpp/io_realm_internal_objectserver_ObjectServerSession.cpp
	realm/realm-library/src/main/cpp/objectserver_shared.hpp
	realm/realm-library/src/main/java/io/realm/internal/SharedRealm.java

+++ realm/realm-library/src/main/cpp/io_realm_RealmFileUserStore.cpp
@@ -91,7 +91,7 @@ Java_io_realm_RealmFileUserStore_nativeGetAllUsers (JNIEnv *env, jclass)
             ThrowException(env, OutOfMemory, ERR_COULD_NOT_ALLOCATE_MEMORY);
             return nullptr;
         }
-        for (unsigned int i = 0; i < len; ++i) {
+        for (size_t i = 0; i < len; ++i) {
             env->SetObjectArrayElement(users_token, i, to_jstring(env, all_users[i]->refresh_token().data()));
         }
